created http methods in controller

+++ src/main/java/com/bridgelabz/addressbookapp/repository/IPersonRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.bridgelabz.addressbookapp.model.PersonData;
 
-public interface IPersonRepository extends JpaRepository<PersonData, Long> {
+public interface IPersonRepository extends JpaRepository<PersonData, Integer> {
 
 }
